+++ Tools/ExcelExport/qtexcel/src/TabelAttr.cpp
@@ -16,26 +16,26 @@ void TableAttr::parseInRange(std::string idrange)
 	int idxFenGeFu = 0;	// "-" 分隔符索引
 	IDRange* pidrange;
 	std::string substring;
-	idxcur = idrange.find_first_of(";", 0);
+	idxcur = (int)idrange.find_first_of(";", 0);
 
 	while(idxcur != -1)
 	{
 		substring = idrange.substr(idxpre, idxcur - idxpre);
 		// 解析子字符串
-		idxFenGeFu = substring.find_first_of("-", 0);
+		idxFenGeFu = (int)substring.find_first_of("-", 0);
 		m_idList.push_back(IDRange());
 		pidrange = &m_idList.back();
 		pidrange->m_startId = atoi(substring.substr(0, idxFenGeFu).c_str());
 		pidrange->m_endId = atoi(substring.substr(idxFenGeFu + 1, substring.length()).c_str());
 
 		idxpre = idxcur + 1;
-		idxcur = idrange.find_first_of(";", idxcur + 1);
+		idxcur = (int)idrange.find_first_of(";", idxcur + 1);
 	}
 
 	// 把最后一个解析了
 	substring = idrange.substr(idxpre, idrange.length());
 
-	idxFenGeFu = substring.find_first_of("-", 0);
+	idxFenGeFu = (int)substring.find_first_of("-", 0);
 	m_idList.push_back(IDRange());
 	pidrange = &m_idList.back();
 	pidrange->m_startId = atoi(substring.substr(0, idxFenGeFu).c_str());
